+++ src/async/territories.js
@@ -6,6 +6,7 @@ import { conn } from '../server';
 import axios from 'axios';
 import addressAsync from './addresses';
 import activityLog from './activityLog';
+import { AddressAsync } from './addresses';
 
 class TerritoryAsync {
   async getTerritory (id) {
@@ -139,6 +140,12 @@ class TerritoryAsync {
 
   async delete (id) {
     if (!id) throw new Error('id is required');
+
+    const addresses = await AddressAsync.getAddressesByTerritory(id);
+    if (addresses.length) {
+      throw new Error('Cannot delete a territory containing addresses');
+    }
+
     const sql = `DELETE FROM territories WHERE id = ${id}`;
     return await conn.query(sql);
   }
